refactor: moved LegendColorMapRow and LegendTypeSwitcher to maplibre folder
