+++ sites/geohub/src/components/maplibre/LegendTypeSwitcher.svelte
